+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## Unreleased
+* Adds `<A> A.validateResult(predicate: (A) -> Boolean, error: (A) -> Throwable): Result<A>` (Andrew Parker)
 
 ## [1.0.2]
 * Moves demand extension function to extensions package (Hugo Müller-Downing)