+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 * Adds `<A> A.validateResult(predicate: (A) -> Boolean, error: (A) -> Throwable): Result<A>` (Andrew Parker)
-* Adds `<A> Result<A>.asOption(): Option<A>` (Andrew Parker)
-* Adds `<A> Result<A>.failureAsOption(): Option<T>` (Andrew Parker)
+* Adds `<A> Result<A>.toOption(): Option<A>` (Andrew Parker)
+* Adds `<A> Result<A>.failureToOption(): Option<T>` (Andrew Parker)
 
 ## [1.0.2]
 * Moves demand extension function to extensions package (Hugo Müller-Downing)